Home

Keys.jl

Keys.False

Keys.Key

Keys.Keyed

Keys.KeyedTuple

Keys.True

Keys.TypedBool

Keys.delete

Keys.fill_tuple

Keys.filter_unrolled

Keys.find_unrolled

Keys.flatten_unrolled

Keys.getindex_unrolled

Keys.if_else

Keys.key

Keys.map_values

Keys.not

Keys.product_unrolled

Keys.push

Keys.reduce_unrolled

Keys.rename

Keys.same_type

Keys.setindex_unrolled

Keys.value

Keys.@_

Keys.@__str

Keys.@q

Keys.@q_

Keys.False — Type.

a TypedBool

source

Keys.Key — Type.

struct Key{K}

A typed key. See @__str for an easy way to create keys. Use to create Keyed values.

source

Keys.Keyed — Type.

struct Keyed{K, V}

an alias for a Key-value pair. a tuple of Keyed values is aliased as a KeyedTuple.

source

Keys.KeyedTuple — Type.

const KeyedTuple

A tuple with only Keyed values. You can index them with Keys or access them with dots. Duplicated keys are allowed; will return the first match.

julia> using Keys

julia> keyed_tuple = (_"a" => 1, _"b" => 2)
(.a => 1, .b => 2)

julia> keyed_tuple.b
2

julia> keyed_tuple[(_"a", _"b")]
(.a => 1, .b => 2)

julia> keyed_tuple.c
ERROR: Key .c not found
[...]

julia> haskey(keyed_tuple, _"b")
True()

julia> merge(keyed_tuple, (_"a" => 4, _"c" => 3))
(.b => 2, .a => 4, .c => 3)

source

Keys.True — Type.

a TypedBool

source

Keys.TypedBool — Type.

abstract TypedBool

typed bools, True and False, can guarantee type stability in cases where constant propogation is not working for Bools.

julia> using Keys

julia> Bool(False())
false

julia> Bool(True())
true

julia> TypedBool(false)
False()

julia> TypedBool(true)
True()

julia> True() & True() & False()
False()

julia> False() & False() & True()
False()

julia> True() | True() | False()
True()

julia> False() | False() | True()
True()

source

Keys.delete — Method.

delete(keyed_tuple::KeyedTuple, keys::Key...)

delete all Keyed values matching Keys in a KeyedTuple.

julia> using Keys

julia> delete((_"a" => 1, _"b" => 2), _"a")
(.b => 2,)

source

Keys.fill_tuple — Method.

julia> using Keys

julia> fill_tuple((1, 'a', 1.0), "a")
("a", "a", "a")

source

Keys.filter_unrolled — Method.

julia> using Keys

julia> filter_unrolled(identity, (True(), False()))
(True(),)

source

Keys.find_unrolled — Method.

julia> using Keys

julia> find_unrolled((true, false, true))
(1, 3)

source

Keys.flatten_unrolled — Method.

julia> using Keys

julia> flatten_unrolled(((1, 2.0), ("c", 4//4)))
(1, 2.0, "c", 1//1)

source

Keys.getindex_unrolled — Method.

julia> using Keys

julia> getindex_unrolled((1, "a", 1.0), (true, false, true))
(1, 1.0)

source

Keys.key — Method.

key(keyed::Keyed)

get the key of a Keyed value.

julia> using Keys

julia> key.((_"a" => 1, _"b" => 2))
(.a, .b)

source

Keys.map_values — Method.

map_values(f, keyed_tuple::KeyedTuple)

map f over the values of a KeyedTuple.

julia> using Keys

julia> map_values(x -> x + 1, (_"a" => 1, _"b" => 2))
(.a => 2, .b => 3)

source

Keys.not — Method.

not(x)

TypedBool aware version of !.

julia> using Keys

julia> not(True())
False()

julia> not(False())
True()

source

Keys.product_unrolled — Method.

julia> using Keys

julia> product_unrolled((1, 2.0), ("c", 4//4))
((1, "c"), (2.0, "c"), (1, 1//1), (2.0, 1//1))

source

Keys.push — Method.

push(keyed_tuple::KeyedTuple, pairs::Keyed...)

push the Keyed values in pairs into the KeyedTuple, replacing common Keys.

julia> using Keys

julia> push((_"a" => 1, _"b" => 2), _"b" => 4, _"c" => 3)
(.a => 1, .b => 4, .c => 3)

source

Keys.reduce_unrolled — Method.

julia> using Keys

julia> reduce_unrolled(&, (true, false, true))
false

source

Keys.rename — Method.

rename(keyed_tuple::KeyedTuple, pairs_of_keys::PairOfKeys...)

for each pair of Keys, where the first key matches in KeyedTuple, it will be replaced by the second.

julia> using Keys

julia> rename((_"a" => 1, _"b" => 2), _"c" => _"a")
(.c => 1, .b => 2)

source

Keys.same_type — Method.

same_type(a, b)

Check whether a and b are the same type; return a TypedBool.

julia> using Keys

julia> same_type(1, 2)
True()

julia> same_type(1, 2.0)
False()

source

Keys.setindex_unrolled — Method.

julia> using Keys

julia> setindex_unrolled(
            (1, "a", 1.0),
            ('a', 1//1),
            (True(), False(), True())
        )
('a', "a", 1//1)

source

Keys.value — Method.

value(key::Keyed)

get the value of a Keyed value.

julia> using Keys

julia> value.((_"a" => 1, _"b" => 2))
(1, 2)

source

Keys.@_ — Macro.

macro _(body::Expr)

Another syntax for anonymous functions. The arguments are inside the body; the first arguments is _, the second argument is __, etc.

julia> using Keys

julia> 1 |> (@_ _ + 1)
2

julia> map((@_ __ - _), (1, 2), (2, 1))
(1, -1)

source

Keys.@__str — Macro.

@__str

make a key

julia> using Keys

julia> _"a"
.a

source

Keys.@q — Macro.

macro q(body::Expr)

Prepare your code for querying. Body should be a function call, with the function ending with the parity. For each argument greater than the parity, anonymize it (using the rules of @_) and also pass along a quoted version.

julia> using Keys

julia> @q 1
ERROR: LoadError: Expecting a function call
[...]

julia> @q f(1)
ERROR: LoadError: Expecting function to end in a digit
[...]

julia> @q f1()
ERROR: LoadError: Expecting at least 1 argument(s)
[...]

julia> call(source1, source2, anonymous, quoted) = anonymous(source1, source2);

julia> @q call2(1, 2, _ + __)
3

source

Keys.@q_ — Macro.

macro q_(body::Expr)

Convenience wrapper for @q and @_

julia> using Keys

julia> call(source1, source2, anonymous, quoted) = anonymous(source1, source2);

julia> result = @q_ call2(_, 2, _ + __);

julia> result(1)
3

source

Keys.if_else — Method.

if_else(switch, new, old)

TypedBool aware version of ifelse.

julia> using Keys

julia> if_else(true, 1, 0)
1

julia> if_else(True(), 1, 0)
1

julia> if_else(False(), 1, 0)
0

source